Dump iOS only IPA with argument --ignore-ios-check

+++ Sources/appdecrypt/dump.swift
@@ -34,6 +34,10 @@ class Dump {
         if targetUrl.hasSuffix("/") {
             targetUrl.removeLast()
         }
+        
+        var ignoreIOSOnlyCheck = false
+        ignoreIOSOnlyCheck = CommandLine.arguments.contains("--ignore-ios-check")
+        
         #if os(iOS)
         if !targetUrl.hasSuffix("/Payload") {
             targetUrl += "/Payload"
@@ -73,9 +77,13 @@ class Dump {
                     let data = pipe.fileHandleForReading.readDataToEndOfFile()
                     if let output = String(data: data, encoding: String.Encoding.utf8) {
                         if output.contains("LC_VERSION_MIN_IPHONEOS") || output.contains("platform 2") {
-                            consoleIO.writeMessage("This app not support dump on M1 Mac. Because machO PLATFORM is IOS!")
-                            do { try fileManager.removeItem(atPath: targetUrl) } catch {}
-                            exit(-10)
+                            if !ignoreIOSOnlyCheck {
+                                consoleIO.writeMessage("This app can't run on Mac M1 ! However, you can decrypt it anyway by adding argument --ignore-ios-check")
+                                do { try fileManager.removeItem(atPath: targetUrl) } catch {}
+                                exit(-10)
+                            } else {
+                                consoleIO.writeMessage("Warning ! The app is will not run on M1 Mac. Decrypting it anyway.")
+                            }
                         }
                     }
                     #endif
